pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "Tidbyt-Powerwall"
description = "Show Tesla Powerwall data on your Tidbyt"
readme = "README.md"
authors = [{name = "Luca Lesinigo", email = "luca@lesinigo.it"}]
dynamic = ["version"]
requires-python = ">=3.10"
keywords = ["tesla", "powerwall", "tesla_powerwall", "tidbyt"]
license = {text = "GNU General Public License v2 (GPLv2)"}
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: No Input/Output (Daemon)",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: GNU General Public License v2 (GPLv2)",
    "Programming Language :: Python :: 3.10",
]
dependencies = [
    "pillow",
    "requests",
    "tesla-powerwall<0.5.0",
]

[project.optional-dependencies]
dev = [
    "build",
    "black",
    "isort",
    "mypy",
    "pylint",
    "types-Pillow",
    "types-requests",
]

[project.scripts]
tidbyt_powerwall = "tidbyt_powerwall:main"

[project.urls]
Homepage = "https://github.com/lesinigo/Tidbyt-Powerwall"
Issues = "https://github.com/lesinigo/Tidbyt-Powerwall/issues"

# [tool.setuptools]
# py-modules = ["tidbyt_powerwall"]

[tool.setuptools.dynamic]
version = {attr = "tidbyt_powerwall.tidbyt_powerwall.__version__"}

[tool.setuptools.package-data]
"tidbyt_powerwall.fonts" = ["*"]

[tool.black]
target-version = ["py310"]
line-length    = 120

[tool.isort]
py_version  = 310
src_paths   = ["src"]
profile     = "black"
line_length = 120

[tool.pylint.main]
fail-under = 10

[tool.pylint.format]
max-line-length = 120

[tool.pylint."messages control"]
enable = ["useless-suppression"]

[tool.mypy]
python_version      = "3.10"
files               = ["src"]
strict              = true
local_partial_types = true             # will be the default under mypy-2.0
